Changing equality check from = to ==
This commit changes the SimpleTalk grammar to use == instead of the
single = for equality checking. It also updates all of the relevant
tests so they pass.

Note that we are maintaining = as a kind of 'assignment' in the single
case of a repeat construct with a given range, ie:
    repeat with myCount = 2 to 25

The equals symbol used there is unique to that rule and we can change
it, perhaps using 'as' instead of '=' or something similar. TBD.