Refactoring undeclared variable lookups to throw an error
Until now, looking up an arbitrary variable name inside a Part's
current execution context dictionary followed the JS pattern,
returning undefined. Undefined behaves oddly in JS and can produce
strange results when, say, attempting to add it to a number.

We have changed the interpreter's variable reading function to now
check for an undefined value when looking up a variable name. Should
the value be undefined, we now throw an error specifying that the
variable has not been declared yet.

Adding a tests to the variable integration test suite for this case.