Fixing SVG Loading issue in FF (see comment)
-- What
There was an issue where SVGs were taking forever to load in FF, and
they were displaying at an enormous size.

The reason for this is a race condition in the rendering logic. It's
unclear if this is a bug in FF or just an error on my part. The crux
is that the SVG has a viewbox set when rendered, but does not have any
SVG width or height attribtues set. These describe the "scale" at
which to render the viewbox. Because we are loading into an element of
undefined size, and because we had a padding set, the renderer was
buggy when trying to determine what to show.

The solution was to check for the presence of width/height attributes
on any new svg element before inserting it into the DOM. If one or the
other of them is not present, we set these values to be equal to
their viewbox equivalents by pulling out that information.

Note that the SVG element width/height attributes are *separate* from
the element.style.width/element.style.height values, which we set in
the style tag to each be 100%. This lets the SVG size up and down if
we adjust the dimensions of the ImageView component manually, as we do
when resizing from Halo.