Adjusting ImageView resize handling and binary img aspect
 -- What
 We have overridden the onHaloResize handler in ImageView to instead
 adjust the dimensions of the underlying image (rather than the
 wrapping component itself). This allows the view to adjust -- "shrink
 wrap" its size to whatever the image is.

We have also implemented a reasonable aspect-ratio-preserving resize
for binary images. It does not track the mouse very well, so we might
want to think of some improvements to that in the future, such as
calculating how far away the cursor actually is from the resize button
at any given moment, slowing down the resizing.