Adding viewport height limit check on loaded images
-- What
Now when a binary or SVG image is loaded onto the page, we calculate
the position of the top left corner and set a max dimension limit to
be that value plus the viewport's corner (minus some padding). This
ensures that when images first load they do not run outside of the viewport.