[BUGFIX] Fixing crucial issue in ExecutionContext re: old context
-- What
Take a look at this example button script:

```
on click
   put "correct" into myVar
   myOtherHandler
   answer myVar
end click

on myOtherHandler
   put (2 + 3) into myVar
end myOtherHandler
```

Prior to this commit, the program would have answered "5" instead of
"correct". This is because we were not restoring a previous execution
context whenever a handler finishes doing whatever it is doing.

I have updated ExecutionContext to track and have the ability to
restore a previous context. You will see that this has been added to
the semantics file at the end of created handler functions.