Adding built in 'tell' command
-- What
The tell sends a command message to some other specified part. This is
a way of sending messages outside of the message delegation chain.

It takes the form of:
   tell <ObjectSpecifier> to <Command>

For example:
    tell first image of this card to loadFromUrl "http://some-image-source"

Note that the message is a deferred one. The tell command's first
argument is the id of the part that should execute it, and the second
argument is a correctly formatted command message object. The target,
once found, will simply send that second argument to itself. This
means that the part invoking the 'tell' command originally *will not
appear in the message's senders list*. This is what we'd expect, since
it is essentially "spoofing" the message send.