Moving stack operations to receiveCommand methods
-- What
Previously on this branch we left the responsibility of creating
handler ActivationContext objects and pushing/popping them from the
ExecutionStack up to the interpreted handlers. This worked but led to
a few potential design issues:
1. Built-in (hardcoded) command handlers would never be a part of the
ExecutionStack;
2. ActivationContexts would not be able to hold information, like
which message triggered the given handler or references to the
concrete JS handler function itself

We have "solved" this my modifying Part.js and System.js
implementations of `receiveCommand` to create the activation for each
handler when it is invoked, push it to the stack, grab the result, pop
the context back off the stack, then return the result.

Note that this required a slight modification to the putInto built-in
handler, which now needs to inspect stackTop - 1 in order to set the
appropriate variables in the context of its caller. We have added the
convenience accessor #previous to the ExecutionStack for this purpose
and also made the change.