[WIP] Fixing properties / serialization issues
-- What
There were several issues regarding Windows and their constituent
Stacks/Cards and how part properties were being
serialized/deserialized.

In short: we create the toolbox and script editors via command
handlers, but we were setting styling and other properties on
them *after* they had been linked to views by setModel(). This allowed
the views to react to the property changes. But on deserialization,
these property changes are not triggered. So instead we updated the
afterModelSet() of the appropriate views to read whatever properties
they would need, like layout for Cards or title for Windows.