[WIP] Serializing current stack / card
-- What
We need to save which card and stack are the current ones into the
serialization. Then, when deserializing, we need to set the
appropriate classes on the relevant views.

In the future, we need to revamp card navigation and current-ness to
be tied to model properties that views respond to, rather than being
solely controlled by the views themselves.