Changing add command to take ObjectSpecifier etc
-- What
Previously, the 'add' command (ie add <systemObject> to <partRef>)
did not take a full ObjectSpecifier as its target in the grammar and
interpreter. This commit resolves that and then some.

-- System.newModel and the newModel command message
This commit required changes to the function signature of the newModel
command handler and message args in System. Previously the function
took a lot of information that is now already handled by the
interpreter when interpreting ObjectSpecifiers, and there is no reason
for it to be part of the handler anymore. We have updated tests and
calls that use this old signature.

Note also that we have added a call to serialize when System receives
a newModel command message, but *not* when the System.newModel()
function is called directly. This is because we do the direct call
when deserializing, and we don't want to call serialize in the middle
of a deserialization cycle because that will overwrite what we are doing.