New Self Hosted Toolbox and Bootstrap
-- What
This commit implements a new bootstrap file that includes a Toolbox
that is completely serialized like any other SimpleTalk part. This
means the Toolbox is no longer a "special" part in the System, but one
that comes pre-serialized into the bootstrap file.

This commit also includes various changes that were necessary to make
this work (some of them concerning serialization/deserialization),
including:
1. Ensuring that new IDs are generated for deserialized parts. We do
not preserve IDs when deserializing anymore, as this causes all kinds
of conflicts. If we ever change our ID schema, we might want to
revisit;
2. Removing the shift-click to activate the Toolbox. This was not
being handled properly and we never noticed it when the Toolbox was a
static special object in the system. For now I have disabled it.
3. Ensuring Windows send the `windowClose` message when the close
button is clicked. The default private command handler is to simply
delete the window model, but scripts on individual windows can
override this. The Toolbox window overrides this and instead sets its
`hide` property to true.
4. Small grammatical and semantic changes
5. Updating stale tests, especially those that checked for the
presence of a Toolbox's stack when initializing an empty System.