Fixing bug with interval id name and ensuring proper deletion
There was a bug where names were not matching for the stored
_stepIntervalId on Parts. This has been resovled and stepping should
work as expected.

Additionally, we ensure that when a Part is deleted from the System,
that we stop all stepping on it, clearing its intervals. If we don't
do this, the intervals will try to run forever and error out repeatedly.