Fixing serialization anchor tag issue
-- What
There was an issue where each time saveHTML is executed and the whole
page HTML is saved, the ephemeral anchor tag used to trigger the save
dialog was also being serialized (but not removed) into the new
document. This meant that, in theory, there would be a new anchor tag
added to the end of the page each time any save was made.

We have fixed this by ensuring that we create/remove the anchor
tag *after* we have saved a snapshot of the HTML state to a
variable. It should be fixed now.