Initial refactored version of list layouts
-- What
the layout and list-direction properties are still categorized as
style props, however due to the complexities of CSS they do not result
in any styler transformations. Instead, they will set/remove specified
CSS classes on the concerned views via the onPropertyChange
callbacks. These are set in PartView. Parts that do not have the
layout and list-direction style properties will never trigger the
change messages, and so the methods in those views are superfluous.

Toolbox is now re-serialized and implemented with a list column layout
on its first card.