[WIP] Adding clipping options to Area and other fixes
-- What
Adding two new properties to Area, clipping and allow-scroll, which
wrap the combined CSS overflow behaviors. These are applied to a
shadow dom slot wrapper so as to never clip the Halo when it is open
on an Area.

Also including a basic example file for toggling layouts in an Area.

Note that clipping is not yet working in a scripted way. Still need to debug.