Fixing issue of current-card deserializing
-- What
Previously, we had CardView check to see if any Card in its parent
stack was set to current and, if not, it would set itself to current
whenever it first gets attached to a StackView. This was throwing off
deserialization, however, when the current card is not the first card
in the serialized stack.

We've gone ahead and removed this auto-current thing from CardView,
and will need to account for it in any loadFromEmpty Systems now.