[FIX] Fixing issues with navigation and current-ness
-- What
In the course of creating the layout example stack, I found that
navigation was buggy and that certain changes I had made were
breaking tests all over the place.

It turns out this has to do with when the WorldStack or any Stack
creates its initial subpart Stack/Card (during live initialization or
deserialization), and when a given stack or card should be set to be
the "current" one.

The primary solution employed in this commit is the one proposed in
Issue #115 -- the "current-ness" of a Stack or Card part should be a
dynamic property (not just a view element class) and the views should
react to that.

In this case, both Stack and Card Parts now have a dynamic property
called 'current' which sets a boolean on the private JS property
_current. When setting to true for any Part, the setter will loop
through sibling Cards/Stacks and unset their current-ness just in
case, ensuring that only one of a kind in a given WorldStack/Stack has
a current true property.

Views have been updated to account for this, as have the navigation tests.