[BUGFIX] Parts on a Stack can now be moved when halo open
-- What
In order to properly manage layouts and ensure that parts inside of an
owner with a list layout cannot me moved (when halo open), we use the
PartView >> #wantsHaloMove getter. This was previously checking for
the value of the owner model's layout property. However, Stacks
don't (for the moment) have layout properties, and thus attempting to
move a Part that was in a Stack triggered a property lookup error.

This has been resolved.