Fixing issue where Halo resizing was adding px to width prop
-- What
We found that the halo was setting width/height values to be
strings (with px at the end) when users resized using the Halo. This
screws up our ability to use arithmetic in scripting when doing things
with widths and heights. Furthermore, the styler already adds the px
as needed behind the scenes when creating the CSS needed to actually
display the correct width/height values on elements.

I have update accordingly.

Additionally, I have added a new Progress Bar example card to the
layout examples stack snapshot