Fixing Halo nested open issues and implementing single halo only
-- What
There was a difference between shift clicking and middle button
clicking when it came to halos -- the middle click would work for the
hierarchical click (deactivate current halo and select parent), but
not when shift clicking. This required rewriting how base click stuff
works between PartView and some of the more specific views.

We are also implementing a single halo at a time functionality. From
now on, only one Halo can be open in the system.