Initial commit of new STClipboard and related features
-- What
The new STClipboard class is designed to be used programmatically for
copying a Part hierarchy, storing it in serializable form, and the
ability to paste it back inside of some target part with all new IDs.

Note that we are trying to make this structure somewhat compatible
with the forthcoming Clipboard API in the browser. For the moment this
doesn't interact at all with any system or browser level
clipboard/copy paste commands. It is its own thing.