Dealing with ambiguity of non-specific add command inside a tell
-- What
We had a confusing situation where the following code, executed in a
field:
tell first card of current stack to add button

Was attempting to add a button to the field itself. This happened for
a couple of reasons:
1. The interpreter was assuming the current partContext as the
target (wrong);
2. We had no way to deal with ambiguity like this inside of a tell
command.

The solution was to rejigger the `newModelCmdHandler` private handler,
which was obviated anyway and wasting lines of code, to deal with a
case where no ownerId is specified. In that case, it attempts to add
a new part of the target type to itself.

This works and all tests are now passing.