Fixing bug where copying a card attempted to set non-ex props
There is no top or left property on a card, but we were trying
to set this on paste. This has been accounted for in a hacky way,
but what we might want to do is have all Parts contain a special
method like `afterPaste` where they can define their own custom
behavior when being pasted.