[BUGFIX] Changins getCurrentCard / getCurrentStackModel to work
For whatever reason (which I still can't determine), tests were
failing when trying to use the System.getCurrentStackModel() methods
that used CSS querySelectors to find the views. When I test in the
browser live this is not a problem, so I'm not sure why the automated
tests are having an issue.

Regardless, this is not the right way to get the concrete part/model
in these cases. After the "current-ness" changes, we now have
convenience getters on World and Stack to get these models. I have
updated the System methods in this commit to simply use those getters.