[WIP] Some initial version working
I still cannot figure out the jittering issue when first displaying
the nav menu.

However, I was able to deal with scaled display oddness in the
wrapped-views by setting and unsetting the display property on the
nested lens stack/card views *after* the drawer has initially come
out. You can see this in action by tracking these three methods:
WrappedView >> #showContent
WrappedView >> #hideContent
StackRow >> #showInitially
CardRow >> #showInitially

The two "showInitially" methods will be called only when the navbar is
opening for the first time. What we should do next is try to make this
promised based using callbacks or to have serious events for when the
nav is opened and closed all the way.

Also the StackRow is not exactly as functional as the CardRow, since
the current stack wrapper is not highlighted nor does it appropriately
respond to click events. I need to test it in a multiple stack scenario.