Initial commit of new STDeserializer
-- What
We are introducing a new STDeserializer object for handling
deserialization. It lays out each step of the process in its main
deserialize() method, which returns a promise for better handling by
the System.

I have also made additional minor changes to other objects:
* Parts now have a special setPropsFromDeserializer method which will
deprecate (eventually) the older setFromDeserialized. Unlike the
former, this also takes the deserializer instance itself as a second
argument.
* The 'compile' message sent to system must now explicity pass a
property called 'serialize' with a 'true' value if the System should
re-serialize itself when compiling. This re-serialization was part of
the reason deserialization was so slow, since we compile a script for
every object that comes in.