Updating how IDs are used and fixing test errors
-- What
This commit adds a flag to the deserializer called
useOriginalIds. It is false by default, meaning anything that is
deserialized will get a new, automatically-generated ID. However, when
true, we make every effort to preserve the ID as specified in the
serialization. This is useful in cases like the initialLoad of a
snapshot.

Additionally, I have updated some tests to account for the new
promise/async nature of initialLoad and deserialization in System.