[BUGFIX] Fixing Nav loading issue on empty pages
-- What
System's initialLoad function was not setting up the navigator in the
correct order in the case where the whole system is loading from
empty (no serialization). That is usually the case with most of our
tests, so they were almost all failing from the uncaught error.

This has now been fixed by ensuring that a World exists *before*
trying to set the Navigator's model to that world.