[WIP] Initial commit of Navigator fixes
-- What
We are using some new patterns. First, CardRows are created and cached
within the Navigator component when models are set. The current stack
then sets the slot name for the correct card row in
handleCurrentChange. This prevents re-creating whole lensed trees and
scaling them each time we toggle through multiple Stacks.

Additionally, we have made WrappedView an actual PartView and told it
how to respond to its model, which will be a Card or Stack.

Other adjustments have been made to deal with these changes.

Things appear to be working well, and snappy. There is still the issue
with the Card that has a layout (layout-examples)