Fixing Field/Nav issue
-- What
Nav was working fine, except in cases where it was hidden and a new
Field was added to the current card. It turns out that FieldView set's
the browser's focus when it is appended. This is problematic for the
Nav, since it's WrappedViews also contain FieldViews. In other words,
the miniature FieldView added to the Nav was forcing itself to come
into focus, leading the Browser to scroll up the contents of the body
so it was in view.

I have removed the autofocus of FieldView for now and it is working.