Fixing queries in tests
-- What
Certain tests were failing because they assumed only a single
hierarchy of stacks and cards. With Navigator there is now an
alternative hierarchy, and it might appear in the DOM *before* the
world. The solution is to include st-world in querySelector statements
where needed so we can be sure we are accessing the primary view
elements in tests.