Make Lenin hand a stack-level object
This means that the hand follows around on different cards as expected.
It also grabs onto objects on the current card and leaves them on the
current card if cards are switched (which sounds confusing, but is what
you would expect).

This does not make the hand a world level object.