Fixing issue where button clicks were not prevented
-- What
When selecting an element for the target, any present onclick handler
was being triggered first before target-pairing. This is due to click
events firing in the order in which they were attached. The solution
is to remove the default click event during the targeting mouseover
and to add it back during targeting mouseleave.