System now sends event when deleteModel removes view(s)
-- What
When we add a model and its corresponding main view, the System
already sends the CustomEvent `st-view-added` to the parent element of
the added main view for the new model. This allows the Navigator to
update its lensed display of <wrapped-view> elements in the Stack and
Card rows, if needed.

This commit adds the same capability for deleteModel. Now we send
`st-view-removed` to the parent element of any removed view. The
Navigator components have been updated to deal with this, including
re-applying their own numerical index display.