Fixing subpart item overflow and adding navigation
-- What
The subpart items list was not scrolling overflow due to complicated
CSS. This has now been fixed.

Additionally, we've added a check to the main Editor's render method
that will navigate to a given Card/Stack if it is not set as the
current one. This way you can see what is actually being edited in the
case of cards and stacks