Styling, command, and test fixes
-- What
This command fixes up some minor styling issues in the editor.

It also fixes the Halo script editor button so that it actually opens
the script editor and not the main editor.

Additionally, we are removing all dead tests and references to
irrelevant objects.

There is no longer an `editorOpen`
property. Should we need that back we will want to put it on the WorldStack.