Adding ID copying to all locations in subparts pane. Adding self.
-- What
This commit adds copy buttons for the id of all referenced subpart
ancestor locations in the Subparts pane of the Editor.

Additionally, we are adding location and id copying buttons for the
part itself in the subparts pane, so that one can see/copy the
location of the part currently being edited.