Fixing issue with subparts location with kinds/ancestors
-- What
There was an error in the code when looking up kind/ancestor pairs for
the location item element. Parts that were directly on a Stack, for
example, still expected to find a Card ancestor even though sometimes
that is not the case (because the part is directly on the stack). Same
goes for Window-ed parts. We now simply hide the item in cases where
no appropriate ancestor is found.