Initial pass as basic open-editor syntax highlighting
-- What
This is an experimental first-pass and syntax highlighting. Script
highlighting will only occur when a script is first opened (see the
openScriptEditor handler in System.js).

The application of styling elements is handled in a new ohm semantics
called AltSyntaxHighlighter in the utils dir.

Styles are applied to elements in the shadow dom of FieldView. I think
we should slot the contents of the editable div so that they can
appear in the live dom, and thus be styled from the stylesheet.

Note that the methods employed here are non sustainable for
comprehensive syntax highlighting (manual addition of spaces, etc).