Ensuring that two st-editors are not in a live System
-- What
An existing st-editor was being serialized with snapshots, and the
initial loading of the System did not check for an existing st-editor
element before creating one. This resulted in the potential for many
editor elements in a given live system/snapshot.

I have updated System to check for the presence of st-editor
elements first, removing them. This ensures that if we update the
editor some old version isn't used.

Additionally, I have updated bootstrap to remove the serialized
existing editor.