Fixing Nav Lensed StackView update issue
-- What
For a long time now, the lensed Stack in the Navigator was not
updating the current card when navigating to another Card, either via
the CardRow in the Navigator or programmatically in the environment.

The issue was related to PR #33, which introduced current-ness as
being set to a specific ID, and the way lensed parts work. The lensed
StackView was not finding any next current card to display, because
the selector used to find it (part-id=) is actually lens-part-id when
a given view is in lensed mode.

I have updated a couple of lines in the handleCurrentChange()
implementation of StackView to adjust for this case and it seems to
work now.