Fixing Nav lensed issue concerning pasted PartView subtrees
-- What
There was a long-standing issue where any pasted Parts that contained
trees of subparts would not have the contents of those trees lensed an
reflected in the Nav view wrappers.

It turns out this was an oversight in the clipboard's implementation
of pasting -- we needed to not only ensure that lensed views of the
pasted item were being created and attached, but that this was done
recursively for any subparts of the newly-pasted item.