Adding submenu capability to ContextMenu
-- What
This commit adds the ability to nest sub-ContextMenus inside of other
ContextMenu items, ie hierarchical menus. To reveal a submenu, simply
hover on an item that has a submenu (indicated by the caret).

We introduce a new component called
ContextMenuItem (st-context-menu-item) which makes this interaction
easier the manage.

We also implement the first such submenu: one that allows authors to
add a new part to the current context of the menu. The list presented
in the submenu only contains subparts that the current context Part
accepts. If it does not accept any parts, we do not even render the
menu item in the first place (ie, with buttons).