Switching ContextMenu to render in document.body
-- What
Previously, the ContextMenu would open in the current
WorldView. However, if the Editor pane was open and the WorldView had
been translated, the ContextMenu would not open in the "correct" (ie
also translated) position on the screen.

The solution is to just render the ContextMenu in the body of the
whole document and avoid any need to translate.