Updating EditorMessenger to accept StatementList and interpret
-- What
Issue #69 pointed out a bug where sending any user-compiled message
from the Editor's Messenger pane would throw a colossal error. This
was due to the fact that we were using a ratchet temporary user
handler called 'doIt' and compiling the statements in this temporary
message handler. The effect of that was to erase all previously
existing user handlers, a big no-no.

Now we match on the grammar and interpret directly from the Part's
existing semantics.