Updating Area's overflow handling; Fixing card 1 home stack
-- What
This commit is the first of several dealing with the "home stack
scrolling" issue.

One aspect of this had to do with how the AreaView element and it's
mapped shadow wrapper element play with each other. It's not good
enough to set the shadow element to overflow: auto; we must also set
the outer AreaView element (host element) to overflow: hidden at the
same time for there to be an effect.

This commit makes those changes.

Additionally, I have updated the properties of areas etc on the first
card of the home stack so that the lower half scrolls properly now.

Something to keep in mind: if we use list-alignment "center" on some
area, but also want it to scroll, the top of that area will scroll up
and out of view equally to the bottom. In order to get normal
scrolling behavior, we need these areas to have list-alignment:
top (when in column view).