Setting empty areas to not shrink at all in flex mode
-- What
When we use Areas as "spacers" in a flex layout, we want to ensure
that they are not "flex-shrinking," which can be the normal behavior
for elements in certain circumstances. Flex shrinking means the
element will shrink proportionally as its parent flex-container
shrinks. In the case of Area spacers, this is exactly what we don't
want.

The new CSS rule in this commit prevents that from happening.