Fixing Image Part Serialization
-- What
Until now, we were always serializing both the `src` and `imageData`
props for all images. The `imageData` property can contain binary
image data that has been loaded. If there is both a `src` (a URL) and
the binary data in the serialization, it's redundant and takes up an
enormous amount of space.

I have rectified this by overriding the Image part's handling of
`serialize()`.

Additionally, I have added a new part property for Image called
`force-serialization`, which, when set to true, will force `imageData`
to be serialized regardless of the presence of any `src` URL.