bug in stack.goToCardById stack reference
the current implementation didn't account for the card being
on a different stack, and would set the target card as "current"
on the context stack. This in turn would trigger the default current
property on the target card's stack as card 1. Effectively any command
like `go to card 2 on stack 2` would actually `go to card 1 of stack 2`