Make pages relocatable
This changes the paths in the examples and index.html files so that they
all refer to the bundle and ohm files as being in the same location. The
files in the `js/objects/examples/` folder were changed by running the
following:

    for file in *.html
    do
        sed -i 's/\.\.\/\.\.\/ohm\/simpletalk\.ohm/\/simpletalk.ohm/g' $file
        sed -i 's/\.\.\/build\/devSystem\.bundle\.js/\/devSystem.bundle.js/g' $file
    done

This commit does not change how the bundle is built, but it does require
it to be served differently. Now you need to serve it in the `serve/` by
executing e.g.

    $ cd serve && python3 -m http.server 8000